+++ packages/core/cairo_alpha/src/tests/with_components_off/erc721/erc721.test.ts
@@ -84,6 +84,22 @@ testERC721('enumerable', {
   enumerable: true,
 });
 
+testERC721('wrapper', {
+  wrapper: true,
+});
+
+test('wrapper adds component and constructor arg', t => {
+  const code = printContract(
+    buildERC721({
+      name: NAME,
+      symbol: SYMBOL,
+      wrapper: true,
+    }),
+  );
+  t.regex(code, /ERC721WrapperComponent/);
+  t.regex(code, /underlying: ContractAddress/);
+});
+
 testERC721('pausable + enumerable', {
   pausable: true,
   enumerable: true,